# specs/001-budget-tracker/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Budget Tracker API
  description: REST API for managing checking account balance, virtual savings, recurring expenses, and pending payments
  version: 1.0.0
  contact:
    name: Budget Tracker Team
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /savings:
    get:
      tags: [Savings]
      summary: List all virtual savings categories
      description: Retrieves all virtual savings categories ordered by creation date
      operationId: listSavings
      responses:
        '200':
          description: Savings list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VirtualSaving'
              example:
                - id: "660e8400-e29b-41d4-a716-446655440001"
                  name: "Emergency Fund"
                  amount: "1000.00"
                  created_at: "2025-11-21T10:05:00Z"
                  updated_at: "2025-11-21T10:05:00Z"
                - id: "660e8400-e29b-41d4-a716-446655440002"
                  name: "Vacation"
                  amount: "500.00"
                  created_at: "2025-11-21T10:06:00Z"
                  updated_at: "2025-11-21T10:06:00Z"
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Savings]
      summary: Create new virtual savings category
      description: Creates a new savings category with name and amount
      operationId: createSaving
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VirtualSavingInput'
      responses:
        '201':
          description: Savings category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualSaving'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
  /savings/{id}:
    put:
      tags: [Savings]
      summary: Update virtual savings category
      description: Updates name and/or amount of an existing savings category
      operationId: updateSaving
      parameters:
        - $ref: '#/components/parameters/SavingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VirtualSavingInput'
      responses:
        '200':
          description: Savings category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualSaving'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      tags: [Savings]
      summary: Delete virtual savings category
      description: Permanently deletes a savings category and releases its amount back to remaining budget
      operationId: deleteSaving
      parameters:
        - $ref: '#/components/parameters/SavingId'
      responses:
        '204':
          description: Savings category deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
    SavingId:
      name: id
      in: path
      required: true
      description: Virtual saving category UUID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    VirtualSaving:
      type: object
      required:
        - id
        - name
        - amount
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Emergency Fund"
        amount:
          type: string
          format: decimal
          description: Amount allocated (must be >= 0)
          example: "1000.00"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    VirtualSavingInput:
      type: object
      required:
        - name
        - amount
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Savings category name
          example: "Emergency Fund"
        amount:
          type: number
          format: decimal
          minimum: 0
          description: Amount to allocate (must be >= 0)
          example: 1000.00
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Amount must be positive"
            details:
              type: object
              description: Additional error context
              additionalProperties: true
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Amount must be a positive number"
              details:
                field: "amount"
                value: -100
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
              details:
                resource_id: "550e8400-e29b-41d4-a716-446655440000"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"